# api/app/Http/Controllers/MetaboliteController.php
<?php
namespace App\Http\Controllers;

use App\Models\Metabolites;
use Illuminate\Http\Request;

class MetaboliteController extends Controller{
	/**
	 * Metabolites controller instance.
	 *
	 * @return void
	 */
	public function __construct(){
	}

	/**
	 * @OA\Get(
	 *	 path="/api/v2/metabolite",
	 *	 operationId="/api/v2/metabolite",
	 *	 tags={""},
	 *	 @OA\Parameter(
	 *		 name="limit",
	 *		 in="query",
	 *		 description="Controls the number of results in the page.",
	 *		 required=false,
	 *		 @OA\Schema(type="integer", default=100)
	 *	 ),
	 *	 @OA\Parameter(
	 *		 name="offset",
	 *		 in="query",
	 *		 description="Determines the offset for the search results. A limit of 200 and offset of 100, will get the third page of 100 results.",
	 *		 required=false,
	 *		 @OA\Schema(type="integer", default=0)
	 *	 ),
	 *	 @OA\Response(
	 *		 response="200",
	 *		 description="Returns list of metabolites registered within system",
	 *		 @OA\JsonContent()
	 *	 ),
	 *	 @OA\Response(
	 *		 response="400",
	 *		 description="Error: Bad request. ",
	 *	 ),
	 * )
	 */
	public function showAllMetabolites(Request $request){
		$this->validate($request, [
			'limit' => 'integer',
			'offset' => 'integer'
		]);
		$limit = $request->input('limit',100);
		$offset = $request->input('offset',0);

		$fullCnt = Metabolites::count();
		$result = Metabolites::skip($offset)->take($limit)->get();
		$result->makeHidden('footprintWkt')->toArray();

		$eor = false;
		$retObj = [
			'offset' => (int)$offset,
			'limit' => (int)$limit,
			'endOfRecords' => $eor,
			'count' => $fullCnt,
			'results' => $result
		];
		return response()->json($retObj);
	}

	/**
	 * @OA\Get(
	 *	 path="/api/v2/metabolite/{identifier}",
	 *	 operationId="/api/v2/metabolite/identifier",
	 *	 tags={""},
	 *	 @OA\Parameter(
	 *		 name="identifier",
	 *		 in="path",
	 *		 description="PK or metabolites record",
	 *		 required=true,
	 *		 @OA\Schema(type="string")
	 *	 ),
	 *	 @OA\Response(
	 *		 response="200",
	 *		 description="Returns metabase on metabolites registered within system with matching ID",
	 *		 @OA\JsonContent()
	 *	 ),
	 *	 @OA\Response(
	 *		 response="400",
	 *		 description="Error: Bad request. Metabolites identifier is required.",
	 *	 ),
	 * )
	 */
	public function showOneMetabolites($id, Request $request){
		$metabolitesObj = Metabolites::find($id);
		if(!$metabolitesObj){
			$metabolitesObj = ['status' =>false, 'error' => 'Unable to locate metabolite based on input identifier'];
		}
		return response()->json($metabolitesObj);
	}
}