# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Symbiota API",
        "contact": {
            "name": "Symbiota Support Hub Team",
            "email": "symbiota@asu.edu"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "../"
        }
    ],
    "paths": {
...
        "/api/v2/metabolite": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/metabolite",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results in the page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the offset for the search results. A limit of 200 and offset of 100, will get the third page of 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of metabolites registered within system",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
        "/api/v2/metabolite/{identifier}": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/metabolite/identifier",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "PK or metabolites record",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns metabase on metabolites registered within system with matching ID",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Metabolites identifier is required."
                    }
                }
            }
        },
        "/api/v2/morphology": {
            "get": {
                "tags": [
                    "Morphology"
                ],
                "operationId": "/api/v2/morphology",
                "parameters": [
                    {
                        "name": "includeStates",
                        "in": "query",
                        "description": "Controls whether character traits are included within output: 0 [default] = do not include state, 1 = include states ",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results in the page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the offset for the search results. A limit of 200 and offset of 100, will get the third page of 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of morphological characters within system",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
        "/api/v2/morphology/{identifier}": {
            "get": {
                "tags": [
                    "Morphology"
                ],
                "operationId": "/api/v2/morphology/identifier",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Identifier (PK = cid) associated with morphological character",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns character and states asscoiated with matching identifier",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Morphological identifier is required."
                    }
                }
            }
        },
        "/api/v2/morphology/{identifier}/attribute": {
            "get": {
                "tags": [
                    "Morphology"
                ],
                "operationId": "/api/v2/morphology/identifier/attribute",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Identifier (PK = cid) associated with morphological character",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cs",
                        "in": "query",
                        "description": "Limit by character State ID (aka cs)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tid",
                        "in": "query",
                        "description": "Limit by taxon PK (aka tid)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results in the page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the offset for the search results. A limit of 200 and offset of 100, will get the third page of 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of attribute codings for a given morphological character",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
...
    }
}